+++ src/vba_edit/word_vba.py
@@ -175,6 +175,9 @@ IMPORTANT:
         help="Enable logging to file. Optional path can be specified (default: vba_edit.log)",
     )
 
+    check_subparser = check_parser.add_subparsers(dest="subcommand", required=False)
+    check_subparser.add_parser("all", help="Check Trust Access to VBA project model of all suported Office applications")
+    
     # Add common arguments to all subparsers (except check command)
     subparser_list = [edit_parser, import_parser, export_parser]
     for subparser in subparser_list:
@@ -293,14 +296,16 @@ def main() -> None:
         # Set up logging first
         setup_logging(verbose=getattr(args, "verbose", False), logfile=getattr(args, "logfile", None))
 
-        # Run 'check' command (Check if Access VBA project model is accessible )
+        # Run 'check' command (Check if VBA project model is accessible )
         if args.command == "check":
+            from vba_edit.utils import check_vba_trust_access
             try:
-                from vba_edit.utils import check_vba_trust_access
-
-                check_vba_trust_access("word")
+                if args.subcommand == "all":
+                    check_vba_trust_access() # Check all supported Office applications
+                else:
+                    check_vba_trust_access("word") # Check MS Word only
             except Exception as e:
-                logger.error(f"Failed to check Trust Access to MS Access VBA project object model: {str(e)}")
+                logger.error(f"Failed to check Trust Access to VBA project object model: {str(e)}")
             sys.exit(0)
         else:
             handle_word_vba_command(args)