+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for macro-enabled MS PowerPoint documents
 - Option to show program's version number and exit added to all cli interfaces (`--version`)
+- `check all` subcommand for cli entry points, which processes all suported MS Office apps in a single call (replaces calling `python -m vba_edit.utils`)
 <!-- ### Changed -->
 <!-- -  -->